requirements.txt
appdirs==1.4.4
astor==0.8.1
black==25.1.0
click==8.1.8
ConfigArgParse==1.7
lxml==5.4.0
mypy_extensions==1.1.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.7
pyshark==0.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
termcolor==3.0.1
tzdata==2025.2
z3-solver==4.13.4.0
